Update the app to be more "manual" and less automatic
The intent here it to make it clearer that we are building a
stack of messages and then we send them to the model in a single
request. When things were happening automatically, it made that
a little difficult to pick up.

Therefore, adding a user message doesn't automatically send the
messages now. A tool execution request doesn't automatically run
that tool.